% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_utils.R
\name{Center}
\alias{Center}
\title{Recapitulate Seurat centering scaled and transformed data}
\usage{
Center(C)
}
\arguments{
\item{C}{countland object}
}
\value{
countland object with slots \code{centered_counts}
}
\description{
Recapitulate Seurat centering scaled and transformed data
}
