% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_cluster.R
\name{Cluster}
\alias{Cluster}
\title{Perform spectral clustering on dot products.}
\usage{
Cluster(C, n_clusters)
}
\arguments{
\item{C}{countland object}

\item{n_clusters}{number of clusters, integer}
}
\value{
countland object with slot \code{embedding}, \code{cluster_labels}
}
\description{
Perform spectral clustering on dot products.
}
