% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_genes.R
\name{CountIndex}
\alias{CountIndex}
\title{Internal function for calculating count index.}
\usage{
CountIndex(lm)
}
\arguments{
\item{lm}{column vector}
}
\value{
count index = largest n where n cells have >= n counts
}
\description{
Internal function for calculating count index.
}
