% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland.R
\name{countland}
\alias{countland}
\title{Initialize a countland object}
\usage{
countland(m, verbose = True)
}
\arguments{
\item{m}{A sparse dgCMatrix of counts}

\item{verbose}{print statements (default=TRUE)}
}
\value{
countland object
}
\description{
Initialize a countland object
}
