% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_cluster.R
\name{Dots}
\alias{Dots}
\title{Calculate pairwise dot products of counts between all cells.}
\usage{
Dots(C)
}
\arguments{
\item{C}{countland object}
}
\value{
countland object with slot \code{dots}
}
\description{
Calculate pairwise dot products of counts between all cells.
}
