% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMA.R
\name{IMA}
\alias{IMA}
\title{run integer matrix approximation}
\usage{
IMA(X, params)
}
\arguments{
\item{X}{observed data matrix}

\item{params}{parameter object}
}
\value{
U, V, and Lambda matrix factors
}
\description{
run integer matrix approximation
}
