% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMA.R
\name{IMA_Compute_Init_Scaled}
\alias{IMA_Compute_Init_Scaled}
\title{rescale if max val is above upper bound}
\usage{
IMA_Compute_Init_Scaled(h, l_bound, u_bound)
}
\arguments{
\item{h}{matrix to be rescaled}

\item{l_bound}{lower bound}

\item{u_bound}{upper bound}
}
\value{
rescaled matrix
}
\description{
rescale if max val is above upper bound
}
