% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMA.R
\name{IMA_init}
\alias{IMA_init}
\title{function to initialize U, V, and Lambda}
\usage{
IMA_init(X, params)
}
\arguments{
\item{X}{observed data matrix}

\item{params}{parameter object}
}
\value{
initialized U, V, and Lambda matrices
}
\description{
function to initialize U, V, and Lambda
}
