% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMA.R
\name{IMA_params}
\alias{IMA_params}
\title{Parameter class for IMA}
\usage{
IMA_params(
  rank,
  u_bounds,
  l_bounds = c(0, 0),
  maxiter = 1e+06,
  stop_crit = 1e-04
)
}
\arguments{
\item{rank}{target number of features in final matrices}

\item{u_bounds}{upper bounds on integers}

\item{l_bounds}{lower bounds on integers (default = c(0,0))}

\item{maxiter}{maximum number of iterations (default = 1000000)}

\item{stop_crit}{criterion of difference at which to stop (default = 0.0001)}
}
\value{
parameter object
}
\description{
Parameter class for IMA
}
