% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMA.R
\name{IMA_Update_Factor}
\alias{IMA_Update_Factor}
\title{Update factor matrix - see SUSTain code}
\usage{
IMA_Update_Factor(M, coeff, mkrp, mode, lambda_, params)
}
\arguments{
\item{M}{matrix to be updated (either U or V)}

\item{coeff}{matrix used in updating algorithm}

\item{mkrp}{matrix used in updating algorithm}

\item{mode}{whether update U or V}

\item{lambda_}{scaling matrix}

\item{params}{parameter object}
}
\value{
updated matrix and scaling factors
}
\description{
Update factor matrix - see SUSTain code
}
