% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_genes.R
\name{listCols}
\alias{listCols}
\title{Split dgCMatrix into column vectors.}
\usage{
listCols(m)
}
\arguments{
\item{m}{dgCMatrix}
}
\value{
list of column vectors, numeric
}
\description{
Split dgCMatrix into column vectors.
}
