% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_utils.R
\name{Log}
\alias{Log}
\title{Recapitulate Seurat log transformation}
\usage{
Log(C)
}
\arguments{
\item{C}{countland object}
}
\value{
countland object with slots \code{log_counts}
}
\description{
Recapitulate Seurat log transformation
}
