% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_utils.R
\name{Normalize}
\alias{Normalize}
\title{Recapitulate Seurat normalization}
\usage{
Normalize(C)
}
\arguments{
\item{C}{countland object}
}
\value{
countland object with slots \code{norm_factor}, \code{norm_counts}
}
\description{
Recapitulate Seurat normalization
}
