% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_cluster.R
\name{PlotClusters}
\alias{PlotClusters}
\title{Plot cells using spectral embedding of dot products.}
\usage{
PlotClusters(C)
}
\arguments{
\item{C}{countland object}
}
\description{
Plot cells using spectral embedding of dot products.
}
