% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_embed.R
\name{PlotEigengap}
\alias{PlotEigengap}
\title{Plots eigenvalues to investigate the optimal number of clusters}
\usage{
PlotEigengap(C)
}
\arguments{
\item{C}{countland object}
}
\value{
generates plot of eigenvalues by number of components
}
\description{
Plots eigenvalues to investigate the optimal number of clusters
}
\examples{
gold_path <- system.file("testdata", package = "countland", mustWork = TRUE)
gold.data <- Seurat::Read10X(data.dir = gold_path)
C <- countland(gold.data)
C <- Dot(C)
C <- Embed(C,n_components=5)
PlotEigengap(C)
}
