% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_rank_reduction.R
\name{PlotIMA}
\alias{PlotIMA}
\title{Plot cells using integer matrix approximation}
\usage{
PlotIMA(C, x = 1, y = 2, subsample = TRUE)
}
\arguments{
\item{C}{countland object}

\item{x}{feature on x-axis, integer (default=1)}

\item{y}{feature on y-axis, integer (default=2)}

\item{subsample}{if TRUE, use subsampled counts (default), otherwise use counts}
}
\description{
Plot cells using integer matrix approximation
}
