% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_cluster.R
\name{PlotMarker}
\alias{PlotMarker}
\title{Plot cell using spectral embedding and display counts in a given gene.}
\usage{
PlotMarker(C, gene_index)
}
\arguments{
\item{C}{countland object}

\item{gene_index}{index value for gene to visualize}
}
\description{
Plot cell using spectral embedding and display counts in a given gene.
}
