% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_subset.R
\name{PrintGeneNumber}
\alias{PrintGeneNumber}
\title{Restore count matrix to original state}
\usage{
PrintGeneNumber(C)
}
\arguments{
\item{C}{countland object}
}
\value{
countland object
}
\description{
Restore count matrix to original state
}
