% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland.R
\name{RemoveEmpty}
\alias{RemoveEmpty}
\title{Internal function to remove empty columns and rows}
\usage{
RemoveEmpty(C)
}
\arguments{
\item{C}{countland object}
}
\value{
countland object, count matrix updated
}
\description{
Internal function to remove empty columns and rows
}
