% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_utils.R
\name{RescaleVariance}
\alias{RescaleVariance}
\title{Recapitulate Seurat scaling to unit variance}
\usage{
RescaleVariance(C)
}
\arguments{
\item{C}{countland object}
}
\value{
countland object with slots \code{scaled_counts}
}
\description{
Recapitulate Seurat scaling to unit variance
}
