% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_subset.R
\name{RestoreCounts}
\alias{RestoreCounts}
\title{Restore count matrix to original state}
\usage{
RestoreCounts(C)
}
\arguments{
\item{C}{countland object}
}
\value{
countland object
}
\description{
Restore count matrix to original state
}
\examples{
gold_path <- system.file("testdata", package = "countland", mustWork = TRUE)
gold.data <- Seurat::Read10X(data.dir = gold_path)
C <- countland(gold.data)
C <- SubsetGenes(C,gene_indices=1:200)
C <- SubsetCells(C,cell_indices=1:50)
C <- RestoreCounts(C)
}
