% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_cluster.R
\name{ScikitManifoldSpectralEmbedding}
\alias{ScikitManifoldSpectralEmbedding}
\title{Recapitulate scikit.manifold.spectral_embedding from python.}
\usage{
ScikitManifoldSpectralEmbedding(A, n_components)
}
\arguments{
\item{A}{similarity matrix, dgCMatrix}

\item{n_components}{number of eigenvectors to retain, integer}
}
\value{
matrix of eigenvectors
}
\description{
Recapitulate scikit.manifold.spectral_embedding from python.
}
