% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_genes.R
\name{Subsample}
\alias{Subsample}
\title{Subsample cells to a standard number of counts by randomly sampling observations without replacement.}
\usage{
Subsample(C, n_counts)
}
\arguments{
\item{C}{countland object}

\item{n_counts}{number of counts to subsample, must be larger than min total counts per cell}
}
\value{
countland object with slot \code{subsample}
}
\description{
Subsample cells to a standard number of counts by randomly sampling observations without replacement.
}
