% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_genes.R
\name{SubsampleCol}
\alias{SubsampleCol}
\title{Internal function for subsampling a column from a sparse matrix.}
\usage{
SubsampleCol(lm, li, j, n_counts)
}
\arguments{
\item{lm}{column vector}

\item{li}{row positions}

\item{j}{column index}

\item{n_counts}{count to sample}
}
\value{
subsampled column as dgTMatrix components
}
\description{
Internal function for subsampling a column from a sparse matrix.
}
