% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland.R
\name{SubsetCells}
\alias{SubsetCells}
\title{Subsets cells using a vecvtor of cell indices}
\usage{
SubsetCells(C, cell_indices)
}
\arguments{
\item{C}{countland object}

\item{cell_indices}{vector of cell index values}
}
\value{
countland object, count matrix updated
}
\description{
Subsets cells using a vecvtor of cell indices
}
