% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland.R
\name{SubsetGenes}
\alias{SubsetGenes}
\title{Subsets genes using a vector of gene indices}
\usage{
SubsetGenes(C, gene_indices)
}
\arguments{
\item{C}{countland object}

\item{gene_indices}{vector of gene index values}
}
\value{
countland object, cout matrix updated
}
\description{
Subsets genes using a vector of gene indices
}
