% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland.R
\name{countland}
\alias{countland}
\title{Initialize a countland object}
\usage{
countland(m, remove_empty = TRUE, verbose = TRUE)
}
\arguments{
\item{m}{A matrix of counts (dense or sparse)}

\item{remove_empty}{filter out cells and genes with no observed counts (default=TRUE)}

\item{verbose}{print statements (default=TRUE)}
}
\value{
countland object
}
\description{
Initialize a countland object
}
\examples{
m <- matrix(sample(seq(0,4),prob=c(0.95,0.3,0.1,0.05,0.05),2000,replace=TRUE),ncol=50)
rownames(m) <- paste0("gene",seq_len(nrow(m)))
colnames(m) <- paste0("cell",seq_len(ncol(m)))
C <- countland(m)
}
