% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_cluster.R
\name{PlotEigengap}
\alias{PlotEigengap}
\title{Plots eigenvalues to investigate the optimal number of clusters}
\usage{
PlotEigengap(C)
}
\arguments{
\item{C}{countland object}
}
\description{
Plots eigenvalues to investigate the optimal number of clusters
}
