% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_cluster.R
\name{PlotEmbedding}
\alias{PlotEmbedding}
\title{Plot cells using spectral embedding of dot products.}
\usage{
PlotEmbedding(C, colors = color_palette)
}
\arguments{
\item{C}{countland object}

\item{colors}{color palette for ggplot2, default=palette of 11 colors}
}
\description{
Plot cells using spectral embedding of dot products.
}
