% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland.R
\name{RestoreCounts}
\alias{RestoreCounts}
\title{Restore count matrix to original state}
\usage{
RestoreCounts(C)
}
\arguments{
\item{C}{countland object}
}
\value{
countland object
}
\description{
Restore count matrix to original state
}
