% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_rank_reduction.R
\name{RunIMA}
\alias{RunIMA}
\title{Perofrm integer matrix approximatin on count matrix.}
\usage{
RunIMA(
  C,
  features,
  u_bounds,
  l_bounds = c(0, 0),
  maxiter = 1e+06,
  stop_crit = 1e-04,
  subsample = TRUE
)
}
\arguments{
\item{C}{countland objecvt}

\item{features}{target number of features, integer}

\item{u_bounds}{upper bounds for U and V matrices, vector of length 2}

\item{l_bounds}{lower bounds for U and V matrices, vector of length 2 (default=c(0,0))}

\item{maxiter}{maximum number of iterations, integer (default=1000000)}

\item{stop_crit}{criterion for stopping based on difference between iterations, numeric (default=0.0001)}

\item{subsample}{if TRUE, use subsampled counts (default), otherwise use counts}
}
\value{
countland object with slots \code{matrixU}, \code{matrixV}, \code{matrixLambda}
}
\description{
Perofrm integer matrix approximatin on count matrix.
}
