% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_genes.R
\name{ScoreGenes}
\alias{ScoreGenes}
\title{Calculate several scores for count-based gene expression.}
\usage{
ScoreGenes(C, subsample = FALSE)
}
\arguments{
\item{C}{countland object}

\item{subsample}{if TRUE, use subsampled counts, otherwise use counts (default=FALSE)}
}
\value{
countland object with slot gene_scores
}
\description{
Calculate several scores for count-based gene expression.
}
