% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countland_genes.R
\name{Subsample}
\alias{Subsample}
\title{Subsample cells to a standard number of counts by randomly sampling observations without replacement.}
\usage{
Subsample(C, gene_counts = NA, cell_counts = NA)
}
\arguments{
\item{C}{countland object}

\item{gene_counts}{maximum total counts for genes}

\item{cell_counts}{sequencing depth for all cells, or "min" to use the minimum cell total}
}
\value{
countland object with slot \code{subsample}
}
\description{
Subsample cells to a standard number of counts by randomly sampling observations without replacement.
}
